pyproject.toml
[project]
name = "comppsych"
version = "0.1.0"
description = "LLM Foraging Task Engine - Adapting cognitive tasks for LLM evaluation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "openai>=1.50.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "seaborn>=0.13.2",
    "streamlit>=1.40.0",
]

[project.scripts]
comppsych = "app:main"
